move cloud types to common
