+++ x-pack/platform/plugins/shared/cloud/server/routes/types.ts
@@ -11,19 +11,3 @@ import { CustomRequestHandlerContext } from '@kbn/core/server';
  * @internal
  */
 export type CloudRequestHandlerContext = CustomRequestHandlerContext<{}>;
-export type SolutionType = 'search' | 'elasticsearch' | 'observability' | 'security';
-export interface CloudDataAttributes {
-  onboardingData: {
-    solutionType?: SolutionType;
-    token: string;
-    security?: CloudSecurityAnswer;
-  };
-}
-
-export interface CloudSecurityAnswer {
-  useCase: 'siem' | 'cloud' | 'edr' | 'other';
-  migration?: {
-    value: boolean;
-    type?: 'splunk' | 'other';
-  };
-}
